Move settings logic from config to utils
